chore: import base components through baseUrl-supported paths
For composite components like Board which refer to subcomponents defined within
the Board-scope we stick to the use of relative path imports such that moving
and renaming of Board will not break the Board component. Any root-level
components are imported through the absolute path as enabled through the
definition of baseUrl.