chore: remove default stories
something is really slow and polluting the env. Not sure what it is but
I noticed that h1 components where being rendered with styles that were
not explicitly defined by Tailwind. From some investigation it appeared
that h1 adopted the styles defined in stories/header.css. I added the
welcome class to an h1 and noticed that the component was rendered with
the properties of the welcome class defined in stories/header.css. Now
we just remove the stories folder to ensure that we don't have anything
leaking into the stories views by accident and tricking us to believe
that something renders a certain way.