refactor: destinationForHeading -> destinationForHeadingId
Reduces the scope of information passed along to the func. We only need the
heading id and the doc context to determine the heading. This should simplify
integration as Headings and ToC components don't always have the full heading
object available to pass along, but they all have the id.